Reorganized the files and renamed files.
